requirements.txt
# Core dependency
PyYAML==6.0.1

# Optional (if you want linting & testing tools)
pytest==8.0.0
flake8==7.0.0




